refactor(tools): merge list-sessions + export into gemini-sessions (#108)
Both tools operated on the same SQLite session store. Collapsing them into
one tool keyed on presence of sessionId removes a redundant schema and gives
hosts a single mental model.

- Create src/tools/gemini-sessions.ts: schema accepts optional sessionId +
  optional format/lastN. No sessionId -> list path (returns sessions, total).
  With sessionId -> export path (returns sessionId, turnCount, format, turns,
  content, exportedAt, optional lastN/totalTurnCount). outputSchema uses
  oneOf to declare both shapes.
- Wire the merged tool into src/dispatcher.ts and src/index.ts ListTools array.
- Delete src/tools/gemini-list-sessions.ts, src/tools/gemini-export.ts,
  tests/tools/gemini-list-sessions.test.ts, tests/tools/gemini-export.test.ts.
- Create tests/tools/gemini-sessions.test.ts covering both paths: list with/
  without sessions, expiresAt computation, dispatcher routing, getTurns
  isolation, all error cases (empty sessionId, missing session, lastN<=0),
  json/markdown formats, lastN windowing, empty-session edge cases.
- README: replace the two tool sections with one gemini-sessions section
  documenting the sessionId discriminator + both output shapes.

Tool surface drops from 9 to 8 tools. Behavior is preserved end-to-end:
old gemini-list-sessions({}) is now gemini-sessions({}); old
gemini-export({sessionId, ...}) is now gemini-sessions({sessionId, ...}).

642 tests pass; npm run build + npx tsc --noEmit clean.
