+++ README.md
@@ -224,36 +224,34 @@ Output:
   cli.detectionError     string | null   Error message if detection failed.
 ```
 
-### `gemini-list-sessions` — list active sessions
+### `gemini-sessions` — list or export sessions
+
+The discriminator is the presence of `sessionId`: omit it to list active
+sessions; provide it to export that session's full conversation history.
 
 ```
 Input:
-  (none)
+  sessionId   string   Optional. When omitted, returns the list of active sessions.
+                       When provided, returns the full conversation history of that session.
+  format      string   Optional. "json" (default) or "markdown". Only used with sessionId.
+  lastN       number   Optional. Export only the last N turns. Only used with sessionId.
 
-Output:
+Output (no sessionId — list path):
   sessions   array    Array of session objects, sorted by most recently accessed.
-    id             string   Session ID (pass to gemini-reply or gemini-export).
-    lastAccessed   string   ISO 8601 timestamp of the last turn.
+    id             string   Session ID (pass back as sessionId to export this session).
+    lastAccessed   number   Unix-ms timestamp of the last turn.
     turnCount      number   Number of turns stored in the session.
-    expiresAt      string   ISO 8601 timestamp when the session will expire (60 min after lastAccessed).
+    expiresAt      number   Unix-ms timestamp when the session will expire (60 min after lastAccessed).
   total      number   Total number of sessions.
-```
 
-### `gemini-export` — export session history
-
-```
-Input:
-  sessionId   string   Required. Session to export.
-  format      string   Optional. "json" (default) or "markdown".
-  lastN       number   Optional. Export only the last N turns (most recent). Omit for all turns.
-
-Output:
+Output (with sessionId — export path):
   sessionId       string   The exported session ID.
   turnCount       number   Number of turns included in the export.
   totalTurnCount  number   Total turns in the session (present only when lastN is used).
   format          string   The format used ("json" or "markdown").
-  turns           array    (format: "json") Array of { role, content } objects.
-  content         string   (format: "markdown") Formatted transcript as a markdown string.
+  turns           array    Array of { role, content } objects.
+  content         string   Pre-rendered representation in the requested format
+                           (JSON.stringify of turns, or bold-label markdown paragraphs).
   exportedAt      string   ISO 8601 timestamp of the export.
 ```
